added matrices topic
