+++ funwithphysics/src/App.js
@@ -49,6 +49,8 @@ import CalcPhotonics from "./Components/Photonics/Topic/Calculator";
 import Contact from "./Components/Contact/Contact";
 import Algebra from "./Components/Algebra/algebra";
 import Geometry from "./Components/Geometry/geometry";
+import Matrices from "./Components/Matrices/matrix";
+import CalMatrix from "./Components/Matrices/Topic/Calculator";
 import Calpermutation from "./Components/Algebra/Topic/Calculator";
 import CalcComplexNumber from "./Components/Algebra/Topic/Calculator";
 import Calstraightline from "./Components/Geometry/Topic/Calculator";
@@ -255,6 +257,16 @@ const App = () => {
           ></Route>
           <Route
             exact
+            path="/matrices"
+            element={
+              <>
+                <Navbar />
+                <Matrices />
+              </>
+            }
+          ></Route>
+          <Route
+            exact
             path="/integration"
             element={
               <>
@@ -288,6 +300,11 @@ const App = () => {
             exact
             path="/geometry/calc/:topic"
             element={<CalCircle />}
+          ></Route>
+          <Route
+            exact
+            path="/matrices/calc/:topic"
+            element={<CalMatrix />}
           ></Route>
           <Route
             exact
