+++ funwithphysics/src/Components/Matrices/Topic/Calculator.css
@@ -1,0 +1,212 @@
+.Calculator__main {
+    margin: auto;
+    padding: 3em;
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-areas:
+        "header header"
+        "details formula"
+        "details process"
+        "example example"
+        "calc calc";
+    column-gap: 3em;
+    row-gap: 1.5em;
+    max-width: 100%;
+}
+
+.K_energy {
+    background-color: white !important;
+}
+
+.Calculator__header {
+    grid-area: header;
+    align-self: center;
+    text-align: center;
+    font-size: 1.8em;
+    color: #e84545;
+    font-weight: 900;
+}
+
+.Calculator__details {
+    grid-area: details;
+    align-self: center;
+    text-align: center;
+    font-size: 1.3em;
+    font-weight: bold;
+    justify-self: center;
+    word-wrap: break-word;
+    width: 30em;
+}
+
+.Calculator__formula {
+    grid-area: formula;
+    font-family: math;
+}
+
+.Calculator__example {
+    grid-area: example;
+    align-self: start;
+    text-align: left;
+    margin-left: 20px;
+}
+
+.Calculator__calc {
+    grid-area: calc;
+}
+
+.Calculator__calc hr {
+    width: auto;
+    border: 1px dashed red;
+}
+
+.Calculator__calc h3 {
+    font-size: 1.5em;
+    font-weight: bold;
+    color: #e84545;
+    text-align: center;
+}
+
+.Calculator__process {
+    grid-area: process;
+    word-wrap: break-word;
+    width: 37.5em;
+    font-size: 1.2rem;
+    padding-right: 2vh;
+}
+
+.binomial_result {
+    width: 55%;
+    margin: auto;
+    padding: 10px 20px;
+    border: 1px solid gainsboro;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.6);
+    word-wrap: break-word;
+}
+
+/* set-matrix modal and tables for determiant calculator */
+.myModal {
+    width: 90vw;
+    position: relative;
+    min-width: 90vw;
+}
+
+table {
+    border: 2px solid;
+}
+
+td {
+    border-bottom: 1px solid black;
+    border-right: 1px solid black;
+    text-align: center;
+    padding: 10px;
+}
+
+.matrix {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+@media only screen and (max-width: 992px) {
+    .Calculator__main {
+        grid-template-columns: 1fr;
+        grid-template-areas: "header header""details details""formula formula ""process process ""example example""calc calc";
+        row-gap: 3em;
+    }
+
+    .Calculator__header,
+    .Calculator__details,
+    .Calculator__formula,
+    .Calculator__calc h3,
+    .Calculator__process {
+        text-align: center;
+        max-width: 100%;
+    }
+
+    .Calculator__example {
+        text-align: center;
+        margin: 0;
+
+    }
+
+    .Calculator__formula h3 {
+        margin-left: -20px;
+    }
+
+    .binomial_result {
+        width: 85%;
+    }
+}
+
+/* Complex number calculator text box group */
+
+.complex-num-group {
+    width: 700px;
+    display: flex;
+    justify-content: space-between;
+    margin: auto;
+}
+
+.group-txt {
+    width: 10%;
+    margin: 0 1%;
+}
+
+
+@media only screen and (max-width: 668px) {
+
+    .complex-num-group {
+        width: fit-content;
+    }
+
+    .group-txt {
+        width: 15%;
+    }
+
+    .binomial_result {
+        width: 95%;
+    }
+}
+
+
+
+@media only screen and (max-width: 668px) {
+
+    .Calculator__header,
+    .Calculator__details,
+    .Calculator__formula,
+    .Calculator__calc h3,
+    .Calculator__process {
+        max-width: 100%;
+    }
+
+    .Calculator__main {
+        padding: 1.5rem;
+    }
+
+    .Calculator__calc hr {
+        max-width: 100%;
+    }
+
+    .Calculator__details,
+    .Calculator__process p {
+        text-align: justify;
+    }
+
+    .Calculator__process {
+        padding: 1rem;
+    }
+
+    input {
+        width: 100%;
+    }
+
+    .Calculator__calc hr {
+        width: 100%;
+        margin: 0px auto 0px auto;
+    }
+
+    .Calculator__calc form {
+        padding: 1rem;
+    }
+}
